refactor: DDD구조에 맞게 패키지 수정
