+++ src/main/java/com/moyeoit/domain/post/application/service/PostService.java
@@ -1,19 +1,19 @@
-package com.moyeoit.domain.post.service;
+package com.moyeoit.domain.post.application.service;
 
-import com.moyeoit.domain.post.controller.request.PostCreateRequest;
-import com.moyeoit.domain.post.controller.response.PopularPostResponse;
-import com.moyeoit.domain.post.controller.response.PostCardResponse;
-import com.moyeoit.domain.post.controller.response.PostDetailInfoResponse;
-import com.moyeoit.domain.post.controller.response.PostLikeResponse;
-import com.moyeoit.domain.post.model.Category;
-import com.moyeoit.domain.post.model.Post;
-import com.moyeoit.domain.post.model.PostImage;
-import com.moyeoit.domain.post.model.PostLike;
-import com.moyeoit.domain.post.model.PostLike.TargetType;
-import com.moyeoit.domain.post.repository.CategoryRepository;
-import com.moyeoit.domain.post.repository.PostImageRepository;
-import com.moyeoit.domain.post.repository.PostLikeRepository;
-import com.moyeoit.domain.post.repository.PostRepository;
+import com.moyeoit.domain.post.presentation.request.PostCreateRequest;
+import com.moyeoit.domain.post.presentation.response.PopularPostResponse;
+import com.moyeoit.domain.post.presentation.response.PostCardResponse;
+import com.moyeoit.domain.post.presentation.response.PostDetailInfoResponse;
+import com.moyeoit.domain.post.presentation.response.PostLikeResponse;
+import com.moyeoit.domain.post.domain.Category;
+import com.moyeoit.domain.post.domain.Post;
+import com.moyeoit.domain.post.domain.PostImage;
+import com.moyeoit.domain.post.domain.PostLike;
+import com.moyeoit.domain.post.domain.PostLike.TargetType;
+import com.moyeoit.domain.post.infra.repository.CategoryRepository;
+import com.moyeoit.domain.post.infra.repository.PostImageRepository;
+import com.moyeoit.domain.post.infra.repository.PostLikeRepository;
+import com.moyeoit.domain.post.infra.repository.PostRepository;
 import com.moyeoit.domain.user.domain.User;
 import com.moyeoit.domain.user.domain.repository.UserRepository;
 import com.moyeoit.global.exception.AppException;
