+++ src/main/java/com/moyeoit/domain/post/infra/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.moyeoit.domain.post.repository;
+package com.moyeoit.domain.post.infra.repository;
 
-import com.moyeoit.domain.post.model.Category;
+import com.moyeoit.domain.post.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category,Long> {
