+++ src/main/java/com/moyeoit/domain/post/infra/repository/PostLikeRepository.java
@@ -1,7 +1,7 @@
-package com.moyeoit.domain.post.repository;
+package com.moyeoit.domain.post.infra.repository;
 
-import com.moyeoit.domain.post.model.PostLike;
-import com.moyeoit.domain.post.model.PostLike.TargetType;
+import com.moyeoit.domain.post.domain.PostLike;
+import com.moyeoit.domain.post.domain.PostLike.TargetType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
